+++ src/PowerAuth/v3/EciesEncryptor.cpp
@@ -155,7 +155,6 @@ EciesClientEncryptor::EciesClientEncryptor(EncryptorParametersPtr& parameters,
     _parameters(std::move(parameters)),
     _secrets(std::move(secrets)),
     _time_service(time_service),
-    _fail_on_nosync_time(true),
     _request_nonce(nonce),
     _time_sync_task(-1)
 {
@@ -177,8 +176,8 @@ EncryptedRequest EciesClientEncryptor::encryptRequest(const ByteRange &data)
     if (!canEncryptRequest()) {
         throw Exception(EC_NotAllowed, "Cannot encrypt request");
     }
-    if (!_time_service->isTimeSynchronized() && _fail_on_nosync_time) {
-        throw Exception(EC_TimeNotSynchronized, "Encryption required time synchronized with server");
+    if (!_time_service->isTimeSynchronized()) {
+        CC7_LOG("WARNING: Time service is not synchronized. Encrypted data may be rejected on the server.");
     }
 
     auto timestamp = _time_service->currentTimeMillis();
@@ -236,15 +235,6 @@ ByteArray EciesClientEncryptor::decryptResponse(const EncryptedResponse &respons
     _time_sync_task = 0.0;
 
     return plaintext;
-}
-
-void EciesClientEncryptor::disableFailWhenTimeIsNotSynchronized()
-{
-#if DEBUG
-    _fail_on_nosync_time = false;
-#else
-    throw Exception(EC_InternalError, "Not implemented in RELEASE build");
-#endif
 }
 
 ByteArray EciesClientEncryptor::getAAD(Timestamp timestamp, const ByteRange & nonce, const ByteRange& ephemeral_key) const